+++ src/V1/Api/Payment/PaymentProcessorInterface.php
@@ -9,5 +9,5 @@ interface PaymentProcessorInterface
     /**
      * @throws PaymentProcessException
      */
-    public function pay(float $price): void;
+    public function pay(int $price): void;
 }
